Add internal network option
Sometimes it is possible, that one or more networks are external and the other(s) not. In that case, we have to be able to differentiate between these networks in docker-compose.yml file. To do that, I have set another variable and condition, and put the external check in that condition.